requirements.txt
autopep8==1.6.0
click==8.1.3
Flask==2.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
python-dotenv==0.20.0
toml==0.10.2
Werkzeug==2.1.2
